backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
supabase==2.5.0
google-generativeai==0.7.2
python-dotenv==1.0.1
pydantic-settings==2.3.4
httpx==0.27.0
python-multipart==0.0.9
pytest==8.2.2
pytest-asyncio==0.23.7

